+++ generic/map.go
@@ -7,7 +7,7 @@ import "github.com/mlange-42/arche/ecs"
 // Create one with [NewMap].
 type Map[T any] struct {
 	id    ecs.ID
-	world ecs.WorldComponents
+	world ecs.IWorld
 }
 
 // NewMap creates a new [Map] for a component type.
@@ -15,7 +15,7 @@ type Map[T any] struct {
 // Map provides a type-safe way to access a component type by entity ID.
 //
 // See also [ecs.World.Get], [ecs.World.Has] and [ecs.World.Set].
-func NewMap[T any](world ecs.WorldComponents) Map[T] {
+func NewMap[T any](world ecs.IWorld) Map[T] {
 	return Map[T]{
 		id:    ecs.ComponentID[T](world),
 		world: world,