+++ generic/resource.go
@@ -7,7 +7,7 @@ import "github.com/mlange-42/arche/ecs"
 // Create one with [NewResource].
 type Resource[T any] struct {
 	id    ecs.ResID
-	world ecs.WorldResources
+	world ecs.IWorld
 }
 
 // NewResource creates a new [Resource] mapper for a resource type.
@@ -16,7 +16,7 @@ type Resource[T any] struct {
 // [Resource] provides a type-safe way to access a world resources.
 //
 // See also [ecs.World.GetResource], [ecs.World.HasResource] and [ecs.World.AddResource].
-func NewResource[T any](world ecs.WorldResources) Resource[T] {
+func NewResource[T any](world ecs.IWorld) Resource[T] {
 	return Resource[T]{
 		id:    ecs.ResourceID[T](world),
 		world: world,